Removed unused functions from VSM.
Move functions only used in tests to RichVariantSampleMatrix defined
in src/test.
