+++ src/test/scala/is/hail/io/ExportVCFSuite.scala
@@ -6,6 +6,7 @@ import is.hail.check.Gen
 import is.hail.check.Prop._
 import is.hail.expr._
 import is.hail.utils._
+import is.hail.testUtils._
 import is.hail.variant.{VSMSubgen, Variant, VariantSampleMatrix}
 import org.testng.annotations.Test
