+++ src/test/scala/is/hail/methods/ProgrammaticAnnotationsSuite.scala
@@ -2,6 +2,7 @@ package is.hail.methods
 
 import is.hail.SparkSuite
 import is.hail.utils._
+import is.hail.testUtils._
 import org.testng.annotations.Test
 
 class ProgrammaticAnnotationsSuite extends SparkSuite {
