+++ src/test/scala/is/hail/stats/InbreedingCoefficientSuite.scala
@@ -4,6 +4,7 @@ import is.hail.SparkSuite
 import is.hail.check.Prop._
 import is.hail.check.Properties
 import is.hail.utils._
+import is.hail.testUtils._
 import is.hail.variant._
 import org.testng.annotations.Test
