+++ src/test/scala/is/hail/testUtils/package.scala
@@ -1,0 +1,10 @@
+package is.hail
+
+import is.hail.utils.RichVariantSampleMatrix
+import is.hail.variant.VariantSampleMatrix
+
+import scala.language.implicitConversions
+
+package object testUtils {
+  implicit def toRichVariantSampleMatrix(vsm: VariantSampleMatrix): RichVariantSampleMatrix = new RichVariantSampleMatrix(vsm)
+}
